feat: move cosign to be a built in verifier
