+++ .github/workflows/publish-dev-assets.yml
@@ -61,7 +61,6 @@ jobs:
           docker buildx create --use
           docker buildx build -f ./httpserver/Dockerfile \
             --platform linux/amd64,linux/arm64,linux/arm/v7 \
-            --build-arg build_cosign=true \
             --build-arg build_sbom=true \
             --build-arg build_licensechecker=true \
             --build-arg build_schemavalidator=true \
