+++ cmd/ratify/main.go
@@ -24,6 +24,7 @@ import (
 	_ "github.com/deislabs/ratify/pkg/policyprovider/configpolicy" // register configpolicy policy provider
 	_ "github.com/deislabs/ratify/pkg/policyprovider/regopolicy"   // register regopolicy policy provider
 	_ "github.com/deislabs/ratify/pkg/referrerstore/oras"          // register oras referrer store
+	_ "github.com/deislabs/ratify/pkg/verifier/cosign"             // register cosign verifier
 	_ "github.com/deislabs/ratify/pkg/verifier/notation"           // register notation verifier
 )
